Removed support for non-copying sends.
This feature was deadlocking on the performance benchmarks. On simply
interactive demos in IPython, it is working fine, but when many
messages are sent/recv in a row (like the benchmark) a deadlock
happens.  For now I am removing this code as it is buggy and is
unlikely that it will make a big difference in the performance.