adjust handling of unrecognized sockopts
In general, sockopts are ints, so allow unrecognized int sockopts to
pass through to libzmq.  If they really are invalid, libzmq will
raise the same EINVAL pyzmq did, but this allows new/experimental
flags to still be used with a pyzmq that has not been updated to be
aware of them.