add threadsafe switch to ZMQStream
wraps any calls to io_loop API in add_callback, which is the only thread safe method on the IOLoop object.  This can introduce delays in the events for code under heavy load, but they should have minimal effect, as the same delays would affect the handling of the events in the same way they affect the registration thereof.

Off by default, so existing code will see no changes.