Fix for an Exception raised when processes exit.
Depending on how an application imports modules, if the reference counts
can lead to `zmq.core.context` getting garbage collected upon process
termination *before* the `Context.__dealloc__` method is called.

The problem is that `__dealloc__` calls `getpid` which was imported by
`zmq.core.context`. At least, it tries to call it, but since the module
was GC'ed, is is `None` and you get the following exception.

    Exception TypeError: "'NoneType' object is not callable" in <zmq.core.context.Context object at 0x28ce140> ignored

This change bails out of the typical `term()` function that is called on
`__dealloc__` if `getpid is None`.