Remove unhelpful GIL releases
We used to release the GIL on all libzmq API calls, but many of these calls are not blocking. Since there is a finite cost to GIL transactions, I have removed the GIL releases that are not on blocking calls.

Simple testing shows a notable improvement in performance for multithreaded apps (simultaneous `while True: send/recv` threads).