cffi: fix linger in Socket.close() and Context.destroy()
This changeset fixes bugs regarding how 'linger' was handled in the cffi
backend. Prior to this fix, running the following example would exit
immediately (as expected) when run under the cython backend, but would hang
under the cffi backend (PYZMQ_BACKEND=cffi).

```
  import zmq

  ctx = zmq.Context()
  s = ctx.socket(zmq.REQ)
  s.connect('tcp://localhost:34567')
  s.send('test')

  s.close(linger=0)      # <- cffi backend hangs here...
  ctx.destroy(linger=0)  # <- ... and here
```

Context.destroy() did not correctly handle the case where linger=0, in which
case linger evaluates to False. The fix is to check specifically for None, not
for any non-True value.

Socket.close() was completely ignoring the 'linger' argument.