Make setup test for "sys/un.h" cross-compile-friendly.
The previous test used compile_and_run but running the test binary on
the build host is not always possible, for instance when
cross-compiling. Furthermore it's not necessary in this case since a
missing "sys/un.h" header would be caught at compile time and the lack
of library support at link time.

Therefore this patch replaces compile_and_run with compile_and_forget
which builds the test program but doesn't run it. compile_and_run is
removed since it's now unused.