setup.py: make cleaning actually work.
Fixes regression from 8aa0dc59970137481d2a0f51e0ccb79dec9695a2 where for
years the --all option was simply ignored, because it walked the tree
then checked a nonexistent variable (which was coincidentally set to
some random value in an unrelated part of the function, thus preventing
a NameError) to see if it was a cythonized .c source.

Without a working clean option, it is difficult to build on python 3.7
which requires regenerating these files.