Make memoryviews created from a Frame writeable with the cython backend
Previously, `Frame.buffer` or `memoryview(Frame)` would result in a
readonly view with the cython backend. This was originally done for
safety since concurrent sends and modifications of a Frame are unsafe.
e.g. if you modify data before a non-copying send is fully done you
would corrupt your message. However, modifying data before the send is
complete is a more general issue and making the view readonly only
protected the case where you're sending data you previously received,
but would not help when sending other arbitrary mutable data. See 1566
for more info.

Making the view writeable allows users to avoid a copy when doing things
like creating a mutable numpy array from a non-copying recv so this is
making a tradeoff between speed and a little bit of safety.

Resolves 1566