Make docstring match type annotation
I *think* the type annotation of `Optional[int]` is correct and the docstring of `float, int` is incorrect.  

The generated API docs use the docstring and so indicate that you can use a float or an int, but then when you use a float, you get type errors in your IDE/mypy/etc.