Call ctx.destroy() in Context.__exit__
ensures cleanup on exit of context instead of hang

more likely the expected behavior (especially on error),
but technically not safe for sockets launched into other threads
(this isn't a very sensible use of `with context`, so I'm okay with that).

Show a ResourceWarning when destroy is going to close the socket.